+++ src/components/AddLink/AddLinkButton.js
@@ -1,5 +1,5 @@
 import classNames from "classnames";
-import styles from "../styles/AddLinkButton.module.css";
+import styles from "./AddLinkButton.module.css";
 
 const AddLinkButton = () => {
   return <button className={classNames(styles.cta, styles.cta_short)}>추가하기</button>;
