+++ src/components/Category/CategoryButtonSkeleton.js
@@ -1,4 +1,4 @@
-import styles from "../styles/CategoryButtonSkeleton.module.css";
+import styles from "./CategoryButtonSkeleton.module.css";
 
 const CategoryButtonSkeleton = () => {
   return <div className={styles.CategoryButtonSkeleton}></div>;
