+++ src/components/FolderInfo/FolderInfo.module.css
