+++ src/components/FolderUtils/CurrentFolder.js
@@ -1,4 +1,4 @@
-import styles from "../styles/CurrentFolder.module.css";
+import styles from "./CurrentFolder.module.css";
 
 const CurrentFolder = ({ children }) => {
   return <h2 className={styles.CurrentFolder}>{children}</h2>;
