+++ src/pages/SharedPage.js
@@ -1,11 +1,11 @@
 import { useCallback, useEffect, useState } from "react";
-import styles from "../styles/SharedPage.module.css";
-import NavBar from "../components/NavBar";
-import Footer from "../components/Footer";
-import FolderInfo from "../components/FolderInfo";
-import Card from "../components/Card";
+import styles from "./SharedPage.module.css";
+import NavBar from "../components/NavBar/NavBar";
+import Footer from "../components/Footer/Footer";
+import FolderInfo from "../components/FolderInfo/FolderInfo";
+import Card from "../components/Card/Card";
 import getFolder from "../api/getFolder";
-import Search from "../components/Search";
+import Search from "../components/Search/Search";
 import useAsync from "../hooks/useAsync";
 
 const SharedPage = () => {
