+++ src/pages/SharedPage.module.css
